refactor google tag manager class
